create sentenceReportContent component
